Fix key duplication
For example, there are four components with todo.id.
If we delete the third todo with id 3, the remaining keys are 1, 2, and 4.
If we then add a new todo, its key will be 4.
This is because the new key is the length of the current todo list (3) + 1.
In other words, the existing key and the newly generated key are duplicates.